streams_manipulation_poc/requirements.txt
fdk>=0.1.44
oci>=2.2.18
kafka-python>=1.4.4
confluent-kafka
certifi
